+++ www/WebInterface/inc/classes/user.class.php
@@ -262,19 +262,19 @@ public static function MakePayment($fromPlayer, $fromPlayerUUID, $toPlayer, $toP
 public static function PaymentQuery($playerName, $playerUUID, $amount){global $config;
   if(toBoolean($config['iConomy']['use'])) {
     $query = "UPDATE `".mysql_san($config['iConomy']['table'])."` SET ".
-             "`balance` = `balance` + ".((float)$amount)." ".
+             "`balance` = `balance` + ".((double)$amount)." ".
              "WHERE LOWER(`username`)='".mysql_san(strtolower($playerName))."' LIMIT 1;";
   } else
   if(toBoolean($config['CC']['use'])) {
     $query = "UPDATE `".mysql_san($config['CC']['prefix'])."_balance` JOIN ".mysql_san($config['CC']['prefix'])."_account ON ".
              mysql_san($config['CC']['prefix'])."_account.id = ".mysql_san($config['CC']['prefix'])."_balance.username_id ".
-             "SET ".mysql_san($config['CC']['prefix'])."_balance.balance = ".mysql_san($config['CC']['prefix'])."_balance.balance + ".((float)$amount)." ".
+             "SET ".mysql_san($config['CC']['prefix'])."_balance.balance = ".mysql_san($config['CC']['prefix'])."_balance.balance + ".((double)$amount)." ".
              "WHERE ".mysql_san($config['CC']['prefix'])."_account.uuid = '".mysql_san($playerUUID)."' AND ".
              "LOWER(".mysql_san($config['CC']['prefix'])."_balance.currency_id) = '".mysql_san(strtolower($config['CC']['currency']))."' ".
              "AND LOWER(".mysql_san($config['CC']['prefix'])."_balance.worldName) = '".mysql_san(strtolower($config['CC']['group']))."' LIMIT 1;";
   } else {
     $query = "UPDATE `".$config['table prefix']."Players` SET ".
-             "`money` = `money` + ".((float)$amount)." ".
+             "`money` = `money` + ".((double)$amount)." ".
              "WHERE `uuid`='".mysql_san($playerUUID)."' LIMIT 1;";
   }
   $result = RunQuery($query, __file__, __line__);
